ml_recommendation_engine/requirements.txt
numpy==1.21.6
pandas==1.3.5
scikit-learn==1.0.2
flask==2.0.3
requests==2.27.1
pyjwt==2.4.0
python-dotenv==0.20.0
gunicorn==20.1.0
pytest==7.0.1
matplotlib==3.5.1
seaborn==0.11.2
joblib==1.1.0
scipy==1.7.3 